refactor: predictRecurrence is a query
