--- src/features/common/recurrence/actions/index.ts
@@ -1,1 +1,0 @@
-export * from './predict-recurrence/predict-recurrence.action';
