+++ src/features/common/recurrence/components/predict-recurrence/predict-recurrence.component.html
@@ -2,8 +2,8 @@
   [formGroup]="predictRecurrenceForm"
   [appLoadAction]="predictRecurrence$"
   #loadAction="loadAction"
-  (actionSuccess)="onPredictRecurrenceActionSuccess($any($event))"
-  (actionError)="onPredictRecurrenceActionError($event)"
+  (actionSuccess)="onPredictRecurrenceQuerySuccess($any($event))"
+  (actionError)="onPredictRecurrenceQueryError($event)"
   (ngSubmit)="onSubmitPredictRecurrence(loadAction.start)">
   <div>
     <label for="query">Répétition:</label>
