+++ src/features/common/recurrence/index.ts
@@ -1,3 +1,3 @@
-export * from './actions';
+export * from './queries';
 export * from './modules/recurrence.service.module';
 export * from './providers';
