--- src/features/common/recurrence/providers/actions/index.ts
@@ -1,1 +1,0 @@
-export * from './predict-recurrence.action.provider';
