+++ src/features/common/recurrence/providers/index.ts
@@ -1,1 +1,1 @@
-export * from './actions';
+export * from './queries';
