+++ src/features/common/recurrence/providers/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './predict-recurrence.query.provider';
