+++ src/features/common/recurrence/queries/index.ts
@@ -1,0 +1,1 @@
+export * from './predict-recurrence/predict-recurrence.query';
